Move test double into its own file
