+++ EurofurenceTests/Application/Components/Dealers/View Model Tests/WhenProducingCategoriesViewModel_DealersViewModelFactoryShould.swift
@@ -110,23 +110,3 @@ class WhenProducingCategoriesViewModel_DealersViewModelFactoryShould: XCTestCase
     }
 
 }
-
-class CapturingDealerCategoryViewModelObserver: DealerCategoryViewModelObserver {
-    
-    enum State {
-        case unset
-        case active
-        case inactive
-    }
-    
-    private(set) var state: State = .unset
-    
-    func categoryDidEnterActiveState() {
-        state = .active
-    }
-    
-    func categoryDidEnterInactiveState() {
-        state = .inactive
-    }
-    
-}
